refactor: modularize ThreeFiberGlobe into reusable components
