--- docs/globe.md
@@ -1,3 +1,0 @@
-https://globe.gl/https://globe.gl/
-http://globe.gl/example/hexed-polygons/
-https://www.redblobgames.com/
